Revert "Log in tests"
