+++ Testing_Asos_WebStore/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="DotNetSeleniumExtras.WaitHelpers" version="3.11.0" targetFramework="net472" />
   <package id="MSTest.TestAdapter" version="2.1.2" targetFramework="net472" />
   <package id="MSTest.TestFramework" version="2.1.2" targetFramework="net472" />
   <package id="Selenium.Support" version="3.141.0" targetFramework="net472" />
